feat: use production and development in templates
