+++ src/services/start/docker.services.ts
@@ -11,11 +11,11 @@ import {
   junoDevConfigExist,
   junoDevConfigFile
 } from '../../configs/juno.dev.config';
-import {JUNO_CONFIG_FILENAME, type JUNO_DEV_CONFIG_FILENAME} from '../../constants/constants';
+import {JUNO_DEV_CONFIG_FILENAME} from '../../constants/constants';
 import {assertDockerRunning, checkDockerVersion} from '../../utils/env.utils';
 import {copyTemplateFile, readTemplateFile} from '../../utils/fs.utils';
 import {confirmAndExit} from '../../utils/prompt.utils';
-import {promptConfigType} from '../init.services';
+import {initConfigNoneInteractive, promptConfigType} from '../init.services';
 
 const TEMPLATE_PATH = '../templates/docker';
 const DESTINATION_PATH = process.cwd();
@@ -64,11 +64,13 @@ const assertJunoDevConfig = async () => {
     `A config file is required for development. Would you like the CLI to create one for you?`
   );
 
-  const config = await buildConfigType('satellite');
+  const {configType, configPath} = await buildConfigType('satellite');
 
-  await copyConfigTemplate({
-    template: JUNO_CONFIG_FILENAME,
-    config
+  await copyTemplateFile({
+    template: `${JUNO_DEV_CONFIG_FILENAME}.${configType}`,
+    sourceFolder: TEMPLATE_PATH,
+    destinationFolder: DESTINATION_PATH,
+    ...(nonNullish(configPath) && {destinationFilename: configPath})
   });
 };
 
@@ -79,30 +81,10 @@ const assertJunoConfig = async () => {
 
   await confirmAndExit(`Your project needs a config file for Juno. Should we create one now?`);
 
-  const config = await buildConfigType('skylab');
-
-  await copyConfigTemplate({
-    template: JUNO_CONFIG_FILENAME,
-    config
-  });
+  await initConfigNoneInteractive();
 };
 
 type ConfigContext = 'skylab' | 'satellite';
-
-const copyConfigTemplate = async ({
-  template,
-  config: {configType, configPath}
-}: {
-  template: typeof JUNO_CONFIG_FILENAME | typeof JUNO_DEV_CONFIG_FILENAME;
-  config: PartialConfigFile;
-}) => {
-  await copyTemplateFile({
-    template: `${template}.${configType}`,
-    sourceFolder: TEMPLATE_PATH,
-    destinationFolder: DESTINATION_PATH,
-    ...(nonNullish(configPath) && {destinationFilename: configPath})
-  });
-};
 
 const buildConfigType = async (context: ConfigContext): Promise<PartialConfigFile> => {
   // We try to automatically detect if we should create a TypeScript or JavaScript (mjs) configuration.
@@ -129,10 +111,10 @@ const assertDockerCompose = async () => {
     message: 'What kind of emulator would you like to run locally?',
     choices: [
       {
-        title: `Full experience — closest to production, including Console, UI, and well known services`,
+        title: `Production-like setup with Console UI and known services`,
         value: `skylab`
       },
-      {title: `Lightweight environment — minimal setup without any UI`, value: `satellite`}
+      {title: `Minimal setup without any UI`, value: `satellite`}
     ]
   });
