+++ templates/init/juno.satellite.predeploy.config.js
@@ -3,7 +3,10 @@ import {defineConfig} from '@junobuild/config';
 /** @type {import('@junobuild/config').JunoConfig} */
 export default defineConfig({
   satellite: {
-    id: '<SATELLITE_ID>',
+    ids: {
+      development: '<DEV_SATELLITE_ID>',
+      production: '<PROD_SATELLITE_ID>'
+    },
     source: '<SOURCE>',
     predeploy: ['<COMMAND> build']
   }
